ACTIN-1673: Move algorithm from medication to treatment
